requirements.txt
# Core
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
authlib==1.3.0

# Utilities
python-dotenv==1.0.1
